requirements.txt
django~=3.1.0

djangorestframework~=3.11.0

dj-rest-auth==1.1.0
django-allauth~=0.42.0

pyyaml==5.3.1
uritemplate==3.0.1
drf-yasg==1.17.1

black
flake8
isort
